+++ sourcegen-generator-kotlin/build.gradle.kts
@@ -7,4 +7,14 @@ dependencies {
     implementation(projects.sourcegenGenerator)
     implementation(libs.managed.kotlinpoet)
     implementation(libs.managed.kotlinpoet.javapoet)
+
+    testImplementation(mnTest.micronaut.test.junit5)
+    testImplementation(libs.junit.jupiter.engine)
+}
+
+tasks.withType(Test::class).configureEach {
+    useJUnit()
+    predictiveSelection {
+        enabled = false
+    }
 }
