+++ test-suite-kotlin/build.gradle.kts
@@ -10,9 +10,6 @@ dependencies {
     ksp(projects.sourcegenGeneratorKotlin)
     ksp(projects.testSuiteCustomGenerators)
     implementation(mn.micronaut.inject.kotlin)
-    implementation(projects.sourcegenGenerator)
-    implementation(projects.sourcegenGeneratorKotlin)
-    implementation(projects.sourcegenModel)
     implementation(projects.sourcegenAnnotations)
     implementation(projects.testSuiteCustomAnnotations)
     testImplementation(mnTest.micronaut.test.junit5)
